Removed Deprecated Screens and unused styles
