+++ src/components/CustomDrawer.jsx
@@ -3,7 +3,6 @@ import {
   View,
   Text,
   ImageBackground,
-  Image,
   TouchableOpacity,
 } from 'react-native';
 import {
