+++ src/components/EditExpenseModal.jsx
@@ -1,15 +1,16 @@
 // ExpenseModal.js
 import React from 'react';
 import { View, Text, Modal, TouchableOpacity, Alert, StyleSheet } from 'react-native';
-import ValidatedTextInput from './ValidatedTextInput';
-import EntypoIcon from 'react-native-vector-icons/Entypo';
 import DatePicker from 'react-native-date-picker';
+import { AppInput } from './AppInput';
 
 
-const NewEditModal = ({ isVisible, onClose, onSave, selectedExpense }) => {
+const EditExpenseModal = ({ isVisible, onClose, onSave, selectedExpense }) => {
   const [concept, setConcept] = React.useState("");
   const [amount, setAmount] = React.useState("");
   const [date, setDate] = React.useState(new Date());
+  const [conceptError, setConceptError] = React.useState(false);
+  const [amountError, setAmountError] = React.useState(false);
   const [open, setOpen] = React.useState(false);
 
   React.useEffect(() => {
@@ -26,12 +27,16 @@ const NewEditModal = ({ isVisible, onClose, onSave, selectedExpense }) => {
 
   const checkConceptError = () => {
     let regex = /^[A-Za-z\d\s]+$/;
-    return !regex.test(concept);
+    let isValid = regex.test(concept);
+    setConceptError(!isValid);
+    return !isValid;
   };
 
   const checkAmountError = () => {
     let regex = /^[\d]{1,12}((\.)(\d){1,2})?$/;
-    return !regex.test(amount);
+    let isValid = regex.test(amount);
+    setAmountError(!isValid);
+    return !isValid;
   };
 
 
@@ -72,38 +77,24 @@ const NewEditModal = ({ isVisible, onClose, onSave, selectedExpense }) => {
         <View style={styles.modalContainer}>
           <Text style={styles.modalTitle}>Edit Expense</Text>
 
-          <ValidatedTextInput
-            label="Concept"
+          <AppInput.Concept 
             value={concept}
             onChangeText={setConcept}
-            validationErrorMessage="Concept may only contain letters or numbers"
-            maxLength={100}
-            hasError={checkConceptError}
-            style={styles.validatedTextInput}
+            onEndEditing={checkConceptError}
+            errorMessage={conceptError? "Concept may only contain letters or numbers" : null}
           />
 
-          <ValidatedTextInput
+          <AppInput.Amount 
             value={amount}
-            label="Amount"
             onChangeText={setAmount}
-            keyboardType="numeric"
-            validationErrorMessage="Amount must be positive and limited to cent precision"
-            hasError={checkAmountError}
-            style={styles.validatedTextInput}
-            maxLength={12}
+            onEndEditing={checkAmountError}
+            errorMessage={amountError? "Amount must be positive and limited to cent precision" : null}
           />
 
-<EntypoIcon.Button
-            onPress={() => {
-              setOpen(true);
-            }}
-            backgroundColor="#ffffff"
-            color="black"
-            name="calendar"
-            style={styles.dateButton}
-          >
-            {date.toDateString()}
-          </EntypoIcon.Button>
+          <AppInput.Date 
+            value={date}
+            onPress={() => setOpen(true)}
+          />
 
           <DatePicker
             modal
@@ -140,6 +131,8 @@ const styles = StyleSheet.create({
 
     width: '100%',
     height: '100%',
+    backgroundColor: 'rgba(0, 0, 0, 0.5)',
+
     justifyContent: 'center',
     alignItems: 'center',
   },
@@ -149,110 +142,16 @@ const styles = StyleSheet.create({
     flexDirection: 'column',
 
     width: '90%',
-    height: '75%',
 
     backgroundColor: 'white',
-    borderWidth: 1,
-    borderColor: 'black',
-    padding: 20,
-  },
-
-  categoryModal: {
-    display: 'flex',
-    flexDirection: 'column',
-    marginTop: '50%',
-    marginLeft: '10%',
-    marginRight: '10%',
-    backgroundColor: '#ADC4C0',
     borderRadius: 10,
     padding: 20,
-
-  },
-
-  categoryListItem: {
-    marginTop: '2%',
-    color: 'black',
-  },
-  
-  categoryCancelButton: {
-    alignSelf: 'center',
   },
 
   modalTitle: {
     fontSize: 20,
     fontWeight: 'bold',
     marginBottom: 10,
-  },
-
-  validatedTextInput: {
-    container: {
-      marginTop: 10,
-    },
-    input: {
-
-    },
-    helperText: {
-
-    }
-  },
-  
-  rnPickerSelect: {
-    inputIOS: {
-      display: 'flex',
-      flexDirection: 'row',
-
-      // width: '100%',
-      height: '100%',
-      backgroundColor: '#a0d406',
-
-      color: 'black',
-      fontSize: 14,
-      fontWeight: 'bold',
-
-    },
-    placeholder: {
-      color: 'black',
-      fontSize: 14,
-      fontWeight: 'bold',
-    },
-    inputAndroid: {
-      display: 'flex',
-      flexDirection: 'row',
-
-      // width: '100%',
-      height: '100%',
-      backgroundColor: '#a0d406',
-
-      color: 'black',
-      fontSize: 14,
-      fontWeight: 'bold',
-    },
-    viewContainer: {
-      display: 'flex',
-      flexDirection: 'row',
-
-      height: 50,
-      width: '100%',
-      marginTop: 40,
-      marginBottom: 40,
-      backgroundColor: '#a0d406',
-
-      justifyContent: 'center',
-      alignItems: 'stretch',
-      borderRadius: 5,
-    }
-  },
-
-  dateButton: {
-    backgroundColor: '#eddeed',
-    height: 50,
-  },  
-
-  input: {
-    backgroundColor: 'white',
-    borderRadius: 5,
-    marginBottom: 10,
-
   },
   saveButton: {
     backgroundColor: '#E86DC3',
@@ -261,8 +160,7 @@ const styles = StyleSheet.create({
     alignItems: 'center',
     marginBottom: 10,
     marginTop: 20,
-    borderColor: 'black',
-    borderWidth: 1, 
+    borderWidth: 1,
   },
   saveButtonText: {
     color: 'white',
@@ -274,8 +172,6 @@ const styles = StyleSheet.create({
     borderRadius: 5,
     padding: 10,
     alignItems: 'center',
-    borderColor: 'black',
-    borderWidth: 1, 
   },
   cancelButtonText: {
     color: 'white',
@@ -284,4 +180,4 @@ const styles = StyleSheet.create({
   },
 });
 
-export default NewEditModal;
+export default EditExpenseModal;
