+++ src/screens/GraphScreen.jsx
@@ -1,19 +1,19 @@
-import React, { useState, useEffect } from 'react';
-import {View, Text, TouchableOpacity, Image, ScrollView, Alert, StyleSheet } from 'react-native';
+import React from 'react';
+import { View, Text, TouchableOpacity, Image, Alert, StyleSheet } from 'react-native';
 import { useNavigation } from '@react-navigation/native';
-import LoadingOverlay from '../components/LoadingOverlay';
-import { BarChart, PieChart,} from "react-native-chart-kit";
+import { PieChart } from "react-native-chart-kit";
 import { fetchUserCategories,fetchExpensesList} from '../utils/apiFetch';
 import { AuthContext } from '../context/AuthContext';
 import FilterModal from '../components/FilterModal';
-import {LinearGradient} from 'react-native-linear-gradient';
+import { LinearGradient } from 'react-native-linear-gradient';
+import { Dialog } from '@rneui/themed';
 
 const GraphScreen = () => {
-  const [loading, setLoading] = useState(false);
-  const [expenses, setExpenses] = useState([]);
-  const [categories, setCategories] = useState([]);
+  const [loading, setLoading] = React.useState(false);
+  const [expenses, setExpenses] = React.useState([]);
+  const [categories, setCategories] = React.useState([]);
   const [isFilterModalVisible, setFilterModalVisible] = React.useState(false);
-  const [legendData, setLegendData] = useState([]);
+  const [legendData, setLegendData] = React.useState([]);
   const {signOut, sessionExpired} = React.useContext(AuthContext);
 
   
@@ -27,7 +27,8 @@ const GraphScreen = () => {
     
   
   const randomColorGenerator = (idx) => {
-    const sliceColor = [      '#DF928E',
+    const sliceColor = [      
+    '#DF928E',
     '#47132D',
     '#CC0066',
     '#A24875',
@@ -48,8 +49,8 @@ const GraphScreen = () => {
     '#894542',
     '#744340',
     '#B47E7B',
-    '#DE9894',]
-    return sliceColor[idx]
+    '#DE9894'];
+    return sliceColor[idx];
   }
 
   const toggleFilterModal = () => {
@@ -119,7 +120,9 @@ const GraphScreen = () => {
     <LinearGradient colors={['#E86DC3', 'white']} style={styles.appContainer}>
 
       <View style={styles.contentContainer}>
-        <LoadingOverlay shown={loading} />
+        <Dialog isVisible={loading}>
+          <Dialog.Loading />
+        </Dialog>
 
         <View style={styles.headerContainer}>
           <View style={styles.logoContainer}>
@@ -211,113 +214,6 @@ const styles = StyleSheet.create({
     resizeMode: 'contain',
   },
 
-  logoutButtonContainer: {
-    display: 'flex',
-    flexDirection: 'row',
-
-    height: '100%',
-    aspectRatio: 0.7,
-
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-
-  logoutButton: {
-    display: 'flex',
-    flexDirection: 'row',
-    borderRadius: 10,
-    backgroundColor: '#d15c54',
-    height: '70%',
-    aspectRatio: 0.7,
-    borderWidth: 1.5,
-    borderColor: "#d15c54",
-    shadowColor: "#000",
-    shadowOffset: {
-      width: 0,
-      height: 2,
-    },
-    shadowOpacity: 0.25,
-    shadowRadius: 3.84,
-
-    elevation: 5,
-
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-
-  logoutButtonIcon: {
-    fontSize: 23,
-  },
-
-
-
-
-  // Menu button container
-
-  menuContainer: {
-    display: 'flex',
-    flexDirection: 'row',
-
-    width: '100%',
-    marginTop: 20,
-    backgroundColor: '#433FF4',
-    justifyContent: 'space-evenly',
-  },
-
-  menuItemContainer: {
-    display: 'flex',
-    flexDirection: 'row',
-
-    width: 100,
-  },
-
-
-
-  rnPickerSelect: {
-    inputIOS: {
-      display: 'flex',
-      flexDirection: 'row',
-
-      // width: '100%',
-      height: '100%',
-      backgroundColor: '#a0d406',
-
-      color: 'black',
-      fontSize: 14,
-      fontWeight: 'bold',
-
-    },
-    placeholder: {
-      color: 'black',
-      fontSize: 14,
-      fontWeight: 'bold',
-    },
-    inputAndroid: {
-      display: 'flex',
-      flexDirection: 'row',
-
-      // width: '100%',
-      height: '100%',
-      backgroundColor: '#a0d406',
-
-      color: 'black',
-      fontSize: 14,
-      fontWeight: 'bold',
-    },
-    viewContainer: {
-      display: 'flex',
-      flexDirection: 'row',
-
-      height: '5%',
-      width: '100%',
-      marginTop: 20,
-      backgroundColor: '#a0d406',
-
-      justifyContent: 'center',
-      alignItems: 'stretch'
-    }
-  },
-
   addExpenseButtonContainer: {
     height: '5%',
     width: '80%',
@@ -334,10 +230,6 @@ const styles = StyleSheet.create({
     alignItems: 'center',
     borderRadius: 10,
   },
-  categorySelectionPicker: {
-    marginTop: '2%',
-    color: 'black',
-  },
   buttonText: {
     color: 'white',
     fontSize: 14,
@@ -345,50 +237,6 @@ const styles = StyleSheet.create({
   },
 
   // Scrollview
-
-  scrollviewContentContainer: {
-    paddingBottom: 20,
-    borderRadius: 20,
-    backgroundColor: 'white', // Background color
-  },
-
-  tableContainer: {
-    display: 'flex',
-    margin: 16,
-    backgroundColor: '#e1e1e8', // Table background color
-    borderWidth: 1,
-    borderColor: '#AEB4E7', // Border color
-    borderRadius: 5,
-  },
-
-  
-  row: {
-    display: 'flex',
-    flexDirection: 'row',
-    width: '100%',
-    height: 40,
-    borderBottomWidth: 1,
-    borderBottomColor: '#ccc',
-  },
-
-  iconContainer: {
-    display: 'flex',
-    flexDirection: 'row',
-
-    height: '100%',
-    aspectRatio: 1,
-    
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-
-  icon: {
-    display: 'flex',
-    flexDirection: 'row',
-    color: 'black',
-
-    fontSize: 25,
-  },
 
   rowMiddleContainer: {
     display: 'flex',
@@ -401,80 +249,6 @@ const styles = StyleSheet.create({
     alignItems: 'stretch',
   },
 
-  conceptContainer: {
-    display: 'flex',
-    flexDirection: 'row',
-
-    width: '100%',
-    height: '50%',
-
-    justifyContent: 'flex-start',
-    alignItems: 'center',
-  },
-
-  concept: {
-    display: 'flex',
-    flexDirection: 'row',
-
-    fontSize: 15,
-    color: '#000000',
-
-    textAlign: 'left',
-  },
-
-  categoryContainer: {
-    display: 'flex',
-    flexDirection: 'row',
-
-    width: '100%',
-    height: '50%',
-
-    justifyContent: 'flex-start',
-    alignItems: 'center',
-  },
-
-  categoryModalContainer: {
-    display: 'flex',
-    flexDirection: 'column',
-    marginTop: '50%',
-    marginLeft: '10%',
-    marginRight: '10%',
-    backgroundColor: '#ADC4C0',
-    borderRadius: 10,
-    padding: 20,
-  },
-
-  category: {
-    display: 'flex',
-    flexDirection: 'row',
-
-    fontSize: 13,
-    color: '#5c5c5c',
-
-    textAlign: 'left',
-  },
-
-  rowLeftContainer: {
-    display: 'flex',
-    flexDirection: 'column',
-
-    width: 80,
-    height: '100%',
-
-    justifyContent: 'flex-start',
-    alignItems: 'stretch',
-  },
-
-  amountContainer: {
-    display: 'flex',
-    flexDirection: 'row',
-
-    width: '100%',
-    height: '50%',
-
-    justifyContent: 'flex-end',
-    alignItems: 'center',
-  },
 
   amount: {
     display: 'flex',
@@ -486,37 +260,11 @@ const styles = StyleSheet.create({
     textAlign: 'left',
   },
 
-  dateContainer: {
-    display: 'flex',
-    flexDirection: 'row',
-
-    width: '100%',
-    height: '50%',
-
-    justifyContent: 'flex-end',
-    alignItems: 'center',
-  },
-
-  date: {
-    display: 'flex',
-    flexDirection: 'row',
-
-    fontSize: 11,
-    color: '#5c5c5c',
-    textAlign: 'left',
-  },
   textForNoData: {
     marginTop: '50%',
     fontSize: 20,
     textAlign: 'center',
   },
-  titulo:{
-    textAlign: 'center',
-    marginTop: 5,
-    fontSize: 18,
-    fontFamily: 'sans-serif-medium',
-
-  }
 });
 
 export default GraphScreen;
